requirements.txt
a2wsgi==1.10.10
aiofiles==25.1.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgi-csrf==0.11
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.3.39
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.8
langchain-core==1.2.9
langchain-ibm==1.0.3
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
lomond==0.3.3
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.4.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.2.3
pathspec==1.0.4
pip==24.0
platformdirs==4.5.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tabulate==0.9.0
tenacity==9.1.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
Werkzeug==3.1.5
xxhash==3.6.0
zstandard==0.25.0

